Ecommerce Menu: AdminMenu separate class added for registering admin menu
